core_runtime_loop/requirements.txt
functions-framework
openai
google-cloud-storage
cloud-sql-python-connector[pg8000]
pg8000
google-auth
